WIP: merging PLDecimal into PLNumber
