+++ src/data/maybe/Maybe.spec.ts
@@ -28,7 +28,7 @@ describe('stdlib/data/Maybe', () => {
 
   describe('toJS', () => {
     it('should return with the JS representation', () => {
-      expect(maybe(plNumber(1)).toJS()).toBe(1)
+      expect(maybe(plNumber(1)).toJS()).toBe({ decimals: 0, intValue: 1 })
       expect(maybe(undefined).toJS()).toBe(undefined)
       expect(maybe(Nothing).toJS()).toBe(undefined)
     })
