+++ src/data/vector/PLVector.spec.ts
@@ -6,6 +6,8 @@ import { plNumber } from '../number/numberFn'
 import { plVector } from './vectorFn'
 import { contains, slice } from '../../typeClasses'
 
+const pln = plNumber
+
 describe('stdlib/data/PLVector', () => {
   describe('creation', () => {
     it('should create empty Vector', () => {
@@ -14,8 +16,8 @@ describe('stdlib/data/PLVector', () => {
     describe('with of', () => {
       it('should have same result as the factory function', () => {
         expect(
-          new PLVector<PLNumber>([plNumber(1), plNumber(2), plNumber(3)]),
-        ).toEqual(plVector(plNumber(1), plNumber(2), plNumber(3)))
+          new PLVector<PLNumber>([pln(1), pln(2), pln(3)]),
+        ).toEqual(plVector(pln(1), pln(2), pln(3)))
       })
     })
   })
@@ -23,42 +25,42 @@ describe('stdlib/data/PLVector', () => {
   describe('getters', () => {
     it('should work', () => {
       expect(plVector().value).toEqual([])
-      expect(plVector(plNumber(1), plNumber(2), plNumber(3)).value).toEqual([plNumber(1), plNumber(2), plNumber(3)])
+      expect(plVector(pln(1), pln(2), pln(3)).value).toEqual([pln(1), pln(2), pln(3)])
     })
   })
 
   describe('count', () => {
     it('should work', () => {
-      expect(plVector().count()).toEqual(plNumber(0))
-      expect(plVector(plNumber(1), plNumber(2), plNumber(3)).count()).toEqual(plNumber(3))
+      expect(plVector().count()).toEqual(pln(0))
+      expect(plVector(pln(1), pln(2), pln(3)).count()).toEqual(pln(3))
     })
   })
 
   describe('concat', () => {
     it('should concatenate 2 vectors', () => {
-      const actual = plVector(plNumber(1))
-        .add(plVector(plNumber(2)))
-        .add(plVector(plNumber(3)))
-      expect(actual).toEqual(plVector(plNumber(1), plNumber(2), plNumber(3)))
+      const actual = plVector(pln(1))
+        .add(plVector(pln(2)))
+        .add(plVector(pln(3)))
+      expect(actual).toEqual(plVector(pln(1), pln(2), pln(3)))
     })
   })
 
   describe('map', () => {
     it('should exec a function on all array item', () => {
-      const actual = plVector(plNumber(1), plNumber(2), plNumber(3)).map((a) => plNumber(a.value * 10))
-      expect(actual).toEqual(plVector(plNumber(10), plNumber(20), plNumber(30)))
+      const actual = plVector(pln(1), pln(2), pln(3)).map((a) => pln(a.value * 10))
+      expect(actual).toEqual(plVector(pln(10), pln(20), pln(30)))
     })
 
     it('should work with empty input', () => {
-      const actual = plVector(...[]).map(() => plNumber(10))
+      const actual = plVector(...[]).map(() => pln(10))
       expect(actual).toEqual(plVector(...[]))
     })
   })
 
   describe('filter', () => {
     it('should should filter out items', () => {
-      const actual = plVector(plNumber(1), plNumber(2), plNumber(3), plNumber(2)).filter((a) => a.equals(plNumber(2)))
-      expect(actual).toEqual(plVector(plNumber(2), plNumber(2)))
+      const actual = plVector(pln(1), pln(2), pln(3), pln(2)).filter((a) => a.equals(pln(2)))
+      expect(actual).toEqual(plVector(pln(2), pln(2)))
     })
 
     it('should work with empty input', () => {
@@ -69,12 +71,12 @@ describe('stdlib/data/PLVector', () => {
 
   it('should have proper toString', () => {
     expect(plVector().toString()).toEqual('[]')
-    expect(plVector(plNumber(1), plNumber(2)).toString()).toEqual('[1,2]')
+    expect(plVector(pln(1), pln(2)).toString()).toEqual('[1,2]')
   })
 
   describe('toJS', () => {
     it('should return with the JS representation', () => {
-      expect(plVector(...[plNumber(1), plNumber(2), plNumber(3)]).toJS()).toEqual([1, 2, 3])
+      expect(plVector(...[pln(1), pln(2), pln(3)]).toJS()).toEqual([pln(1), pln(2), pln(3)])
     })
   })
 
@@ -91,8 +93,8 @@ describe('stdlib/data/PLVector', () => {
       const copiedValue = originalValue.copy()
       expect(originalValue).not.toBe(copiedValue)
       expect(originalValue.value).toEqual(copiedValue.value)
-      expect(originalItem).toBe(copiedValue.index(plNumber(0)))
-      expect(originalItem.value).toBe(copiedValue.index(plNumber(0)).value)
+      expect(originalItem).toBe(copiedValue.index(pln(0)))
+      expect(originalItem.value).toBe(copiedValue.index(pln(0)).value)
     })
   })
 
@@ -103,26 +105,26 @@ describe('stdlib/data/PLVector', () => {
       const copiedValue = originalValue.deepCopy()
       expect(originalValue).not.toBe(copiedValue)
       expect(originalValue.value).not.toBe(copiedValue.value)
-      expect(originalItem).not.toBe(copiedValue.index(plNumber(0)))
-      expect(originalItem.value).toBe(copiedValue.index(plNumber(0)).value)
+      expect(originalItem).not.toBe(copiedValue.index(pln(0)))
+      expect(originalItem.value).toBe(copiedValue.index(pln(0)).value)
     })
   })
 
   describe('contains', () => {
     it('should returns with true/false when the item contains or not the item', () => {
-      expect(contains(plNumber(1), plVector(plNumber(1), plNumber(2)))).toEqual(plBool(true))
-      expect(contains(plNumber(5), plVector(plNumber(1), plNumber(2)))).toEqual(plBool(false))
+      expect(contains(pln(1), plVector(pln(1), pln(2)))).toEqual(plBool(true))
+      expect(contains(pln(5), plVector(pln(1), pln(2)))).toEqual(plBool(false))
     })
   })
 
   describe('slice', () => {
     it('should return with a sub vector', () => {
-      const v = plVector(plNumber(1), plNumber(2), plNumber(3), plNumber(4), plNumber(5), plNumber(6))
+      const v = plVector(pln(1), pln(2), pln(3), pln(4), pln(5), pln(6))
 
-      expect(slice(plNumber(0), plNumber(2), v)).toEqual(plVector(plNumber(1), plNumber(2)))
-      expect(slice(plNumber(2), plNumber(2), v)).toEqual(plVector())
-      expect(slice(plNumber(2), plNumber(4), v)).toEqual(plVector(plNumber(3), plNumber(4)))
-      expect(slice(plNumber(5), plNumber(10), v)).toEqual(plVector(plNumber(6)))
+      expect(slice(pln(0), pln(2), v)).toEqual(plVector(pln(1), pln(2)))
+      expect(slice(pln(2), pln(2), v)).toEqual(plVector())
+      expect(slice(pln(2), pln(4), v)).toEqual(plVector(pln(3), pln(4)))
+      expect(slice(pln(5), pln(10), v)).toEqual(plVector(pln(6)))
     })
   })
 })
