+++ src/utils/math.spec.ts
@@ -15,11 +15,11 @@ describe('math utils', () => {
       expect(isBelowEpsilon(-Infinity)).toBe(false)
       expect(isBelowEpsilon(-42)).toBe(false)
       expect(isBelowEpsilon(-1)).toBe(false)
-      expect(isBelowEpsilon(-Number.EPSILON)).toBe(false)
+      expect(isBelowEpsilon(-Number.EPSILON)).toBe(true)
       expect(isBelowEpsilon(-Math.sin(Math.PI))).toBe(true)
       expect(isBelowEpsilon(0)).toBe(true)
       expect(isBelowEpsilon(Math.sin(Math.PI))).toBe(true)
-      expect(isBelowEpsilon(Number.EPSILON)).toBe(false)
+      expect(isBelowEpsilon(Number.EPSILON)).toBe(true)
       expect(isBelowEpsilon(1)).toBe(false)
       expect(isBelowEpsilon(42)).toBe(false)
       expect(isBelowEpsilon(Infinity)).toBe(false)
